Fix heatmap click handling for result categories
This still involves a hack, sadly, as what ought to be the correct
approach doesn't work due to a bug in Dash/Plotly. But it's a slightly
more robust hack than the one we used before so hopefully it won't cause
further issues.

Closes #45
